Resolve more TODO items
