+++ pennylane/templates/subroutines/permute.py
@@ -94,7 +94,7 @@ class Permute(Operation):
             op = qml.Permute([4, 2, 0, 1, 3], wires=[0, 1, 2, 3, 4])
             tape = qml.tape.QuantumTape([op])
 
-        >>> tape_expanded = qml.tape.tape.expand_tape(tape)
+        >>> tape_expanded = qml.tape.expand_tape(tape)
         >>> print(qml.drawer.tape_text(tape_expanded, wire_order=range(5)))
         0: ─╭SWAP───────────────────┤
         1: ─│─────╭SWAP─────────────┤
