+++ pennylane/transforms/zx/converter.py
@@ -383,7 +383,7 @@ def _to_zx_transform(
 
         # Expand the tape to be compatible with PyZX and add rotations first for measurements
         stop_crit = qml.BooleanFn(lambda obj: isinstance(obj, Operator) and obj.name in gate_types)
-        mapped_tape = qml.tape.tape.expand_tape(
+        mapped_tape = qml.tape.expand_tape(
             mapped_tape, depth=10, stop_at=stop_crit, expand_measurements=expand_measurements
         )
